web/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-filter==22.1
django-queryable-properties==1.8.2
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
pbr==5.9.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.28.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
testresources==2.0.1
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
wheel==0.37.1
zipp==3.8.0
gunicorn==20.1.0
whitenoise==3.3.1